refactor: split widget integrations
