+++ src/components/Widget/index.tsx
@@ -1,28 +1,14 @@
-import { TransactionReceipt } from '@ethersproject/abstract-provider'
-import { Percent } from '@uniswap/sdk-core'
-import {
-  Currency,
-  Field,
-  Slippage,
-  SwapController,
-  SwapEventHandlers,
-  SwapSettingsController,
-  SwapWidget,
-  TradeType,
-} from '@uniswap/widgets'
+import { Currency, SwapWidget } from '@uniswap/widgets'
 import { useWeb3React } from '@web3-react/core'
-import CurrencySearchModal from 'components/SearchModal/CurrencySearchModal'
-import { DEFAULT_DEADLINE_FROM_NOW } from 'constants/misc'
 import { RPC_URLS } from 'constants/networks'
 import { useActiveLocale } from 'hooks/useActiveLocale'
-import { useCallback, useMemo, useState } from 'react'
-import {
-  useIsDarkMode,
-  useSetUserSlippageTolerance,
-  useUserSlippageTolerance,
-  useUserTransactionTTL,
-} from 'state/user/hooks'
+import { useMemo } from 'react'
+import { useIsDarkMode } from 'state/user/hooks'
 import { DARK_THEME, LIGHT_THEME } from 'theme/widget'
+
+import { useSyncWidgetSettings } from './settings'
+import { useSyncWidgetTransactions } from './transactions'
+import { useSyncWidgetValue } from './value'
 
 export const WIDGET_WIDTH = 320
 
@@ -59,125 +45,4 @@ export default function Widget({ defaultToken }: WidgetProps) {
       {tokenSelector}
     </>
   )
-}
-
-// Integrates the Widget's settings, keeping the widget and app settings in sync.
-function useSyncWidgetSettings() {
-  const [userTtl, setUserTtl] = useUserTransactionTTL()
-  const [ttl, setTtl] = useState<number | undefined>(userTtl / 60)
-  const onTransactionDeadlineChange = useCallback(
-    (ttl: number | undefined) => {
-      setTtl(ttl)
-      const userTtl = ttl === undefined ? ttl : ttl * 60
-      setUserTtl(userTtl ?? DEFAULT_DEADLINE_FROM_NOW)
-    },
-    [setUserTtl]
-  )
-  const [userSlippage, setUserSlippage] = [useUserSlippageTolerance(), useSetUserSlippageTolerance()]
-  const [slippage, setSlippage] = useState<string | undefined>(
-    userSlippage === 'auto' ? undefined : userSlippage.toFixed(2)
-  )
-  const onSlippageChange = useCallback(
-    (slippage: Slippage) => {
-      setSlippage(slippage.max)
-      if (slippage.auto || !slippage.max) {
-        setUserSlippage('auto')
-      } else {
-        setUserSlippage(new Percent(Math.floor(Number(slippage.max) * 100), 10_000))
-      }
-    },
-    [setUserSlippage]
-  )
-  const onSettingsReset = useCallback(() => {
-    setTtl(undefined)
-    setUserTtl(DEFAULT_DEADLINE_FROM_NOW)
-    setSlippage(undefined)
-    setUserSlippage('auto')
-  }, [setUserSlippage, setUserTtl])
-  const settings: SwapSettingsController = useMemo(() => {
-    const auto = userSlippage === 'auto'
-    return { slippage: { auto, max: slippage }, transactionTtl: ttl }
-  }, [slippage, ttl, userSlippage])
-  const settingsHandlers: SwapEventHandlers = useMemo(
-    () => ({ onSettingsReset, onSlippageChange, onTransactionDeadlineChange }),
-    [onSettingsReset, onSlippageChange, onTransactionDeadlineChange]
-  )
-
-  return { settings: { settings, ...settingsHandlers } }
-}
-
-// Integrates the Widget's controlled value, using the app's CurrencySearchModal for token selection.
-function useSyncWidgetValue(defaultToken: Currency) {
-  const [isExactInput, setIsExactInput] = useState(false)
-  const [amount, setAmount] = useState<string>()
-  const onAmountChange = useCallback((field: Field, amount: string) => {
-    setIsExactInput(field === Field.INPUT)
-    setAmount(amount)
-  }, [])
-  const [tokens, setTokens] = useState<{ [Field.INPUT]?: Currency; [Field.OUTPUT]?: Currency }>({
-    [Field.OUTPUT]: defaultToken,
-  })
-  const onSwitchTokens = useCallback(() => {
-    setIsExactInput((isExactInput) => !isExactInput)
-    setTokens((tokens) => ({
-      [Field.INPUT]: tokens[Field.OUTPUT],
-      [Field.OUTPUT]: tokens[Field.INPUT],
-    }))
-  }, [])
-  const [selectingField, setSelectingField] = useState<Field>()
-  const otherField = useMemo(() => (selectingField === Field.INPUT ? Field.OUTPUT : Field.INPUT), [selectingField])
-  const [selectingToken, otherToken] = useMemo(() => {
-    if (selectingField === undefined) return [undefined, undefined]
-    return [tokens[selectingField], tokens[otherField]]
-  }, [otherField, selectingField, tokens])
-  const onTokenSelectorClick = useCallback((field: Field) => {
-    setSelectingField(field)
-    return false
-  }, [])
-  const onTokenSelect = useCallback(
-    (token: Currency) => {
-      if (selectingField === undefined) return
-      setIsExactInput(selectingField === Field.INPUT)
-      setTokens(() => {
-        return {
-          [otherField]: token === otherToken ? selectingToken : otherToken,
-          [selectingField]: token,
-        }
-      })
-    },
-    [otherField, otherToken, selectingField, selectingToken]
-  )
-  const value: SwapController = useMemo(
-    () => ({ type: isExactInput ? TradeType.EXACT_INPUT : TradeType.EXACT_OUTPUT, amount, ...tokens }),
-    [amount, isExactInput, tokens]
-  )
-  const valueHandlers: SwapEventHandlers = useMemo(
-    () => ({ onAmountChange, onSwitchTokens, onTokenSelectorClick }),
-    [onAmountChange, onSwitchTokens, onTokenSelectorClick]
-  )
-
-  const tokenSelector = (
-    <CurrencySearchModal
-      isOpen={selectingField !== undefined}
-      onDismiss={() => setSelectingField(undefined)}
-      selectedCurrency={selectingToken}
-      otherSelectedCurrency={otherToken}
-      onCurrencySelect={onTokenSelect}
-    />
-  )
-  return { value: { value, ...valueHandlers }, tokenSelector }
-}
-
-// Integrates the Widget's transactions, showing the widget's transactions in the app.
-function useSyncWidgetTransactions() {
-  const txHandlers /*: TransactionEventHandlers */ = useMemo(
-    () => ({
-      onTxSubmit: (hash: string, tx: unknown) => console.log('onTxSubmit'),
-      onTxSuccess: (hash: string, receipt: TransactionReceipt) => console.log('onTxSuccess'),
-      onTxFail: (hash: string, receipt: TransactionReceipt) => console.log('onTxFail'),
-    }),
-    []
-  )
-
-  return { transactions: { ...txHandlers } }
 }
