+++ src/components/Widget/transactions.ts
@@ -1,0 +1,17 @@
+import { TransactionReceipt } from '@ethersproject/abstract-provider'
+import { TransactionEventHandlers } from '@uniswap/widgets'
+import { useMemo } from 'react'
+
+// Integrates the Widget's transactions, showing the widget's transactions in the app.
+export function useSyncWidgetTransactions() {
+  const txHandlers: TransactionEventHandlers = useMemo(
+    () => ({
+      onTxSubmit: (hash: string, tx: unknown) => console.log('onTxSubmit'),
+      onTxSuccess: (hash: string, receipt: TransactionReceipt) => console.log('onTxSuccess'),
+      onTxFail: (hash: string, receipt: TransactionReceipt) => console.log('onTxFail'),
+    }),
+    []
+  )
+
+  return { transactions: { ...txHandlers } }
+}
